% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{return_class_specific_attr}
\alias{return_class_specific_attr}
\title{Display levels for factors, unique, for characters
and range for numeric values}
\usage{
return_class_specific_attr(data)
}
\arguments{
\item{data}{input dataset}
}
\description{
Display levels for factors, unique, for characters
and range for numeric values
}
