% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sum_columns}
\alias{sum_columns}
\title{Sum numeric columns of input data}
\usage{
sum_columns(data)
}
\arguments{
\item{data}{dataset on which to perform operation}
}
\description{
Sum numeric columns of input data
}
